+++ CHANGELOG.md
@@ -24,6 +24,7 @@ CHANGELOG for FlatCAM beta
 - updated the Italian translation (by Massimiliano Golfetto)
 - fixed a series of issues in Gerber Editor tools when the user is trying to use the tools by preselecting a aperture without size (aperture macro)
 - moved all the UI stuff out of the Gerber Editor class in its own class
+- in the Excellon Editor, added shortcut keys Space and Ctrl+Space for toggling the direction of the Slots, respectively for the Array of Slots
 
 2.11.2020